Add function for numbers array generating.

+++ programming-1/problem-3/problem-3/comparator.h
@@ -16,7 +16,7 @@ typedef struct benchmark_result {
 } benchmark_res;
 
 double count_average(double *runtime, size_t num);
-double *run_sort(char sort, size_t calls_number, void *base, size_t num, size_t size, double(*compar)(void*, void*), void*(*prepare_data)(void*, int));
+double *run_sort(char sort, size_t calls_number, void *base, size_t num, size_t size, double(*compar)(void*, void*), void*(*prepare_data)(void*, size_t));
 double count_standard_dev_1(double *runtime, double average, size_t num);
 double count_standard_dev_2(double *runtime, double average, size_t num);
 benchmark_res benchmark(char sort, size_t calls_number, void *base, size_t num, size_t size, double(*compar)(void*, void*), void*(*prepare_data)(void*, int));
